Split out processFile into its own package for gometalinter integration.
